fix for gcc6

+++ es-core/src/Renderer_draw_gl.cpp
@@ -87,7 +87,7 @@ namespace Renderer {
 
 	void drawRect(float x, float y, float w, float h, unsigned int color, GLenum blend_sfactor, GLenum blend_dfactor)
 	{
-		drawRect((int)round(x), (int)round(y), (int)round(w), (int)round(h), color, blend_sfactor, blend_dfactor);
+		drawRect((int)round((long) x), (int)round((long) y), (int)round((long) w), (int)round((long) h), color, blend_sfactor, blend_dfactor);
 	}
 
 	void drawRect(int x, int y, int w, int h, unsigned int color, GLenum blend_sfactor, GLenum blend_dfactor)
